[neutron] Differentiate alerting form networks not synced
We should alert early when a network is not scheduled to any network
agent since DNS and DHCP missing can have bad impact on the ability to
do networky things. On the other side, we can give reconciliation a bit
more time if a network is only missing on one agent, albeit if the
agent is currently reloaded.